Add files via upload
To do this, first, the Forward function of the Convolution layer is implemented in the form of normal multiplication and multiplication based on Stochastic calculations, and at the same time, matrices for Input and Weight are given to both of these functions, and MSE (Mean Square Error) is calculated compared to normal multiplication. And this has been done for 500 matrices, and all these matrices have been entered into the Excel environment at the same time, based on the written codes, and the results have been calculated and the MSE has been calculated.